requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.3.2
langchain==0.1.12
langchain-community==0.0.29
langchain-google-genai==0.0.7
chromadb==0.4.22
tiktoken==0.5.2
pydantic==2.6.1
rank-bm25==0.2.2
pandas==2.1.1
openpyxl==3.1.2
scikit-learn==1.3.0